+++ packages/editor-tools/tests/messages-of-author.test.ts
@@ -1,9 +1,24 @@
 import { test, expect } from "vitest";
-import { MessageOfAuthorGetter, TransformResult } from "../src/messages-of-author";
+import { CustomError, MessageOfAuthor, MessageOfAuthorGetter } from "../src/messages-of-author";
 import { SamplingParams, LmEditorState } from "../src/types";
 
-const getMessagesOfAuthor = (editorState: LmEditorState): TransformResult => {
-  const state = new MessageOfAuthorGetter(editorState);
+type TransformSuccess = {
+  tag: "success";
+  value: MessageOfAuthor[];
+};
+
+type TransformError = {
+  tag: "error";
+  value: CustomError[];
+};
+
+type TransformResult = TransformSuccess | TransformError;
+
+const getMessagesOfAuthor = (
+  editorState: LmEditorState,
+  useGenerationUuids = true,
+): TransformResult => {
+  const state = new MessageOfAuthorGetter(editorState, useGenerationUuids);
 
   const errors = state.getErrors();
   if (errors.length > 0) {
@@ -48,7 +63,7 @@ test("empty state", async () => {
       "value": [
         {
           "author": "user",
-          "parts": [],
+          "tasks": [],
         },
       ],
     }
@@ -82,9 +97,9 @@ test("handles paragraph", async () => {
       "value": [
         {
           "author": "user",
-          "parts": [
+          "tasks": [
             {
-              "tag": "text",
+              "tag": "AddTextTask",
               "text": "What is the best subject for the illustration to accompany the following?",
             },
           ],
@@ -117,9 +132,9 @@ test("handles heading", async () => {
       "value": [
         {
           "author": "user",
-          "parts": [
+          "tasks": [
             {
-              "tag": "text",
+              "tag": "AddTextTask",
               "text": "### Content",
             },
           ],
@@ -164,9 +179,9 @@ test("handles variableSelect", async () => {
       "value": [
         {
           "author": "user",
-          "parts": [
+          "tasks": [
             {
-              "tag": "text",
+              "tag": "AddTextTask",
               "text": ""Question: "What is the person doing?" Answer: "The person is happy""",
             },
           ],
@@ -240,27 +255,87 @@ test("handles lmGenerator", async () => {
       "value": [
         {
           "author": "user",
-          "parts": [
+          "tasks": [
             {
-              "tag": "text",
+              "tag": "AddTextTask",
               "text": "Explanation: ",
             },
             {
-              "nodeAttrs": {
-                "choices": [],
-                "id": "720ddbc0-12e6-4583-83b6-d0229a60445b",
-                "max_tokens": 256,
-                "name": "_explanation",
-                "stop": [
-                  "
+              "max_tokens": 256,
+              "name": "720ddbc0-12e6-4583-83b6-d0229a60445b",
+              "regex": undefined,
+              "stop": [
+                "
     ",
-                ],
-                "type": "generation",
-              },
-              "tag": "lmGenerate",
+              ],
+              "tag": "GenerateTask",
             },
             {
-              "tag": "text",
+              "tag": "AddTextTask",
+              "text": "",
+            },
+          ],
+        },
+      ],
+    }
+  `);
+});
+
+test("handles lmGenerator, use names", async () => {
+  const msgs = getMessagesOfAuthor(
+    {
+      version: "1",
+      variables: [],
+      samplingParams: SAMPLING_PARAMS,
+      doc: {
+        type: "doc",
+        content: [
+          { type: "authorSelect", attrs: { author: "user" } },
+          {
+            type: "paragraph",
+            content: [
+              { type: "text", text: "Explanation: " },
+              {
+                type: "lmGenerator",
+                attrs: {
+                  id: "720ddbc0-12e6-4583-83b6-d0229a60445b",
+                  choices: [],
+                  type: "generation",
+                  max_tokens: 256,
+                  name: "_explanation",
+                  stop: ["\n"],
+                },
+              },
+            ],
+          },
+        ],
+      },
+    },
+    false,
+  );
+  expect(msgs).toMatchInlineSnapshot(`
+    {
+      "tag": "success",
+      "value": [
+        {
+          "author": "user",
+          "tasks": [
+            {
+              "tag": "AddTextTask",
+              "text": "Explanation: ",
+            },
+            {
+              "max_tokens": 256,
+              "name": "_explanation",
+              "regex": undefined,
+              "stop": [
+                "
+    ",
+              ],
+              "tag": "GenerateTask",
+            },
+            {
+              "tag": "AddTextTask",
               "text": "",
             },
           ],