+++ packages/client/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @lmscript/client
 
+## 0.0.32
+
+### Patch Changes
+
+- Improve editor-tools
+
 ## 0.0.31
 
 ### Patch Changes